chore: move `file-download` to `@crawlee/http`
