--- packages/file-download/.npmignore
@@ -1,6 +1,0 @@
-node_modules
-src
-test
-coverage
-apify_storage
-tsconfig.*
