--- packages/file-download/README.md
@@ -1,41 +1,0 @@
-# `@crawlee/file-download`
-
-Provides a framework for the parallel file download using plain HTTP requests. The URLs to download are fed either from a static list of URLs or can be added dynamically from another source.
-
-Since `FileDownload` uses raw HTTP requests to download the files, it is very fast and efficient on data bandwidth. However, it doesn't parse the content - if you need to e.g. extract data from the downloaded files, you might need to use [CheerioCrawler](https://crawlee.dev/api/cheerio-crawler/class/CheerioCrawler), [PuppeteerCrawler](https://crawlee.dev/api/puppeteer-crawler/class/PuppeteerCrawler) or [PlaywrightCrawler](https://crawlee.dev/api/playwright-crawler/class/PlaywrightCrawler) instead.
-
-`FileCrawler` downloads each URL using a plain HTTP request and then invokes the user-provided [FileCrawlerOptions.requestHandler](https://crawlee.dev/api/file-crawler/interface/FileCrawlerOptions#requestHandler) where the user can specify what to do with the downloaded data.
-
-The source URLs are represented using [Request](https://crawlee.dev/api/core/class/Request) objects that are fed from [RequestList](https://crawlee.dev/api/core/class/RequestList) or [RequestQueue](https://crawlee.dev/api/core/class/RequestQueue) instances provided by the [FileCrawlerOptions.requestList](https://crawlee.dev/api/file-crawler/interface/FileCrawlerOptions#requestList) or [FileCrawlerOptions.requestQueue](https://crawlee.dev/api/file-crawler/interface/FileCrawlerOptions#requestQueue) constructor options, respectively.
-
-If both [FileCrawlerOptions.requestList](https://crawlee.dev/api/file-crawler/interface/FileCrawlerOptions#requestList) and [FileCrawlerOptions.requestQueue](https://crawlee.dev/api/file-crawler/interface/FileCrawlerOptions#requestQueue) are used, the instance first processes URLs from the [RequestList](https://crawlee.dev/api/core/class/RequestList) and automatically enqueues all of them to [RequestQueue](https://crawlee.dev/api/core/class/RequestQueue) before it starts their processing. This ensures that a single URL is not crawled multiple times.
-
-The crawler finishes when there are no more [Request](https://crawlee.dev/api/core/class/Request) objects to crawl.
-
-We can use the `preNavigationHooks` to adjust `gotOptions`:
-
-```
-preNavigationHooks: [
-    (crawlingContext, gotOptions) => {
-        // ...
-    },
-]
-```
-
-New requests are only dispatched when there is enough free CPU and memory available, using the functionality provided by the [AutoscaledPool](https://crawlee.dev/api/core/class/AutoscaledPool) class. All [AutoscaledPool](https://crawlee.dev/api/core/class/AutoscaledPool) configuration options can be passed to the `autoscaledPoolOptions` parameter of the `FileCrawler` constructor. For user convenience, the `minConcurrency` and `maxConcurrency` [AutoscaledPool](https://crawlee.dev/api/core/class/AutoscaledPool) options are available directly in the `FileCrawler` constructor.
-
-## Example usage
-
-```ts
-const crawler = new FileDownloader({
-    requestHandler({ body, request }) {
-        writeFileSync(request.url.replace(/[^a-z0-9\.]/gi, '_'), body);
-    },
-});
-
-await crawler.run([
-    'http://www.example.com/document.pdf',
-    'http://www.example.com/sound.mp3',
-    'http://www.example.com/video.mkv',
-]);
-```
