--- packages/file-download/package.json
@@ -1,63 +1,0 @@
-{
-    "name": "@crawlee/file-download",
-    "version": "3.9.2",
-    "description": "The scalable web crawling and scraping library for JavaScript/Node.js. Enables development of data extraction and web automation jobs (not only) with headless Chrome and Puppeteer.",
-    "engines": {
-        "node": ">=16.0.0"
-    },
-    "main": "./dist/index.js",
-    "module": "./dist/index.mjs",
-    "types": "./dist/index.d.ts",
-    "exports": {
-        ".": {
-            "import": "./dist/index.mjs",
-            "require": "./dist/index.js",
-            "types": "./dist/index.d.ts"
-        },
-        "./package.json": "./package.json"
-    },
-    "keywords": [
-        "apify",
-        "headless",
-        "chrome",
-        "puppeteer",
-        "crawler",
-        "scraper"
-    ],
-    "author": {
-        "name": "Apify",
-        "email": "support@apify.com",
-        "url": "https://apify.com"
-    },
-    "contributors": [
-        "Jan Curn <jan@apify.com>",
-        "Marek Trunkat <marek@apify.com>",
-        "Ondra Urban <ondra@apify.com>"
-    ],
-    "license": "Apache-2.0",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/apify/crawlee"
-    },
-    "bugs": {
-        "url": "https://github.com/apify/crawlee/issues"
-    },
-    "homepage": "https://crawlee.dev",
-    "scripts": {
-        "build": "yarn clean && yarn compile && yarn copy",
-        "clean": "rimraf ./dist",
-        "compile": "tsc -p tsconfig.build.json && gen-esm-wrapper ./dist/index.js ./dist/index.mjs",
-        "copy": "tsx ../../scripts/copy.ts"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "dependencies": {
-        "@crawlee/http": "3.9.2",
-        "@crawlee/types": "3.9.2",
-        "@crawlee/utils": "3.9.2",
-        "cheerio": "^1.0.0-rc.12",
-        "htmlparser2": "^9.0.0",
-        "tslib": "^2.4.0"
-    }
-}
