--- packages/file-download/src/index.ts
@@ -1,2 +1,0 @@
-export * from '@crawlee/http';
-export * from './internals/file-download';
