+++ packages/http-crawler/src/index.ts
@@ -1,2 +1,3 @@
 export * from '@crawlee/basic';
 export * from './internals/http-crawler';
+export * from './internals/file-download';
